move env variables
