--- ProgramParsing/ENV_VARIABLES/parse_years.py
@@ -1,9 +1,0 @@
-"""
-parse_years.py is a library built to receive information on Major Requirements
-
-Contributors:
-Brent Huang
-"""
-
-PARSE_YEAR_BEG = 2019
-PARSE_YEAR_END = 2021
