+++ ProgramParsing/ProgramParser/ENV_VARIABLES/parse_years.py
@@ -1,0 +1,16 @@
+"""
+parse_years.py is a library built to receive information on Major Requirements
+
+Contributors:
+Brent Huang
+"""
+
+PARSE_YEAR_BEG = 2019
+PARSE_YEAR_END = 2021
+
+# Must be in this format
+CALENDAR_YEARS = ["2019-2020", "2020-2021", "2021-2022"]
+DEFAULT_YEAR = "2020-2021" # Refers to dict
+
+# For debuggin purposes:
+# CALENDAR_YEARS = ["2020-2021"]
