+++ ProgramParsing/ProgramParser/ParseProgram.py
@@ -1,13 +1,8 @@
 from Database.DatabaseSender import DatabaseSender
-from ProgramParsing.ENV_VARIABLES.parse_years import PARSE_YEAR_BEG, PARSE_YEAR_END
+from ProgramParsing.ProgramParser.ENV_VARIABLES.parse_years import PARSE_YEAR_BEG, PARSE_YEAR_END, CALENDAR_YEARS, DEFAULT_YEAR
 import re
 
-# Must be in this format
-CALENDAR_YEARS = ["2019-2020", "2020-2021", "2021-2022"]
-DEFAULT_YEAR = "2020-2021"
 
-# For debuggin purposes:
-# CALENDAR_YEARS = ["2020-2021"]
 
 def filterFiles(files, filesToIgnore):
     # adds the academic year in front of strings in filesToIgnore
